Refactor toolbar button classes
