+++ src/builtin_plugins/plugin/amulet/editor/layout/_layout.py
@@ -2,7 +2,7 @@
 Registering layouts, adding layout buttons and enabling layouts"""
 
 from __future__ import annotations
-from typing import Callable, cast
+from typing import Callable, cast, TYPE_CHECKING
 from threading import Lock, current_thread, main_thread
 from dataclasses import dataclass
 import re
@@ -10,7 +10,8 @@ from weakref import ref
 
 from PySide6.QtCore import Qt, QPoint, QSize, QObject
 
-from plugin.amulet.editor._icon import ATooltipIconButton
+if TYPE_CHECKING:
+    from plugin.amulet.editor._toolbar_button import ToolbarButton
 
 from plugin.amulet.editor.widget import _missing
 from plugin.amulet.editor.widget import _widget
@@ -78,8 +79,8 @@ class LayoutContainer:
     layout_id: str
     default_config: LayoutConfig
     layout_config: LayoutConfig
-    button_ref: Callable[[], ATooltipIconButton | None] = cast(
-        Callable[[], ATooltipIconButton | None], lambda: None
+    button_ref: Callable[[], ToolbarButton | None] = cast(
+        Callable[[], ToolbarButton | None], lambda: None
     )
     hidden_layout: HiddenLayout | None = None
